requirements.txt


SpeechRecognition
sounddevice
wheel
pyaudio
fastapi
uvicorn[standard]
six
opencv-python
langchain
langchain_openai
langchain_core
langchain_community
rank_bm25
mediapipe
chromadb
facenet_pytorch
openai-whisper